Git LFS cannot push venues.bson file correctly, therefore it is deleted
Git LFS cannot push venues.bson file correctly and it makes some changes in a way after git clone cannot restore the file into  mongodb. As a result users should move the datasets to the related folder before building the an image. we will find a solution for uploading the .bson files into git in a feature